paginationPage.rb: fix jekyll4 cache mismatch errors
Jekyll 4.0 introduced a new caching mechanism to the LiquidRenderer
class. This remembers the paths of any files passed to convert and
the rendered output. The cache is reset between repeat invocations
of `Site.process' it's really only to prevent needing to re render
templates or includes.

The cache works by remembering the file name of any files being
rendered, returning the cached value on demand should jekyll
try to render a file with the same path (from the src directory)
multiple times.

The issue here is that Jekyll-Paginate-V2 sets the name of all
`Jekyll::PaginateV2::Generator::PaginationPage` instances to
index.html. Meaning every pagination page gets rendered to the result
of the first paginated page.

This patch simply assigns some fields the Page instance, such that it
appears to have the same file name as the file to which it's written
to (which should guarantee it's uniqueness and prevent liquid cache
mismatches).